add events to the user model

+++ controllers/event.js
@@ -1,4 +1,6 @@
 const Event = require("../models/eventModel");
+const User = require("../models/userModel");
+const multer = require("multer");
 
 exports.findAllEvents = async (req, res) => {
   try {
@@ -20,32 +22,65 @@ exports.findOneEvent = async (req, res) => {
     res.status(500).json({ message: err.message });
   }
 };
-
 exports.createEvent = async (req, res) => {
   try {
-    console.log(req.body);
+    const { title, description, date, location } = req.body;
+    const eventImg = req.file.path;
+
+    // eventOrganizer and eventOrganizerId are derived from the authenticated user
+    const eventOrganizer = req.user.name;
+    const eventOrganizerId = req.user._id;
 
     const event = new Event({
-      title: req.body.title,
-      description: req.body.description,
-      date: req.body.date,
-      location: req.body.location,
-      eventImg: req.file.path,
+      title,
+      description,
+      date,
+      location,
+      eventImg,
+      eventOrganizer,
+      eventOrganizerId,
     });
 
+    let user = await User.findById(req.user._id);
+    if (!user) {
+      return res.status(404).json({ message: "User not found" });
+    }
+
+    user.events.push(event);
+    await user.save();
+
     const newEvent = await event.save();
-    res.status(201).json(newEvent);
+
+    res.status(201).json({
+      status: "success",
+      newEvent,
+    });
   } catch (err) {
+    console.error(err); // Log the error for debugging
     res.status(400).json({ message: err.message });
   }
 };
 
 exports.deleteEvent = async (req, res) => {
   try {
-    const event = await Event.findById(req.params.id);
+    const event = await Event.findById(
+      req.params.id,
+      (eventOrganizerId = req.user._id)
+    );
     if (!event) {
       return res.status(404).json({ message: "Event not found" });
     }
+    if (post.authorId.toString() !== req.user.id) {
+      return res.status(401).json({
+        status: "Fail",
+        message: `You can only delete a post you created!`,
+      });
+    }
+
+    const user = await User.findById(req.user._id);
+    user.events.pull(event);
+    await user.save();
+
     await event.deleteOne();
     res.status(200).json({ message: "Event deleted successfully" });
   } catch (err) {
@@ -67,6 +102,16 @@ exports.updateEvent = async (req, res) => {
     event.createdBy = req.body.createdBy;
     await event.save();
     res.status(200).json(event);
+  } catch (err) {
+    res.status(500).json({ message: err.message });
+  }
+};
+
+exports.findEventsByUser = async (req, res) => {
+  try {
+    const userId = req.params.userId;
+    const events = await Event.find({ eventOrganizerId: userId });
+    res.status(200).json(events);
   } catch (err) {
     res.status(500).json({ message: err.message });
   }
